requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
branca==0.4.2
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==5.1.1
Django==3.2
django-google-maps==0.12.4
django-phonenumber-field==6.1.0
executing==0.8.3
folium==0.11.0
idna==3.3
ipython==8.1.1
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
numpy==1.22.3
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.12.44
pickleshare==0.7.5
Pillow==6.2.0
prompt-toolkit==3.0.28
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2018.5
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
twilio==6.41.0
urllib3==1.26.8
wcwidth==0.2.5
